# docs/swagger.yaml
basePath: /api/v1
definitions:
  connection.CheckConnectionRequest:
    properties:
      ID:
        type: integer
    required:
    - ID
    type: object
  connection.ConnectionInfoResponse:
    properties:
      UUID:
        type: string
      config:
        $ref: '#/definitions/models.ConnectionConfig'
      createdAt:
        example: "2025-08-22T12:00:00Z"
        type: string
      description:
        type: string
      lastUsed:
        example: "2025-08-22T12:05:00Z"
        type: string
      machineID:
        example: 1
        type: integer
      sessionID:
        example: ns=3;i=3093118269
        type: string
      status:
        $ref: '#/definitions/connection.ConnectionStatusEnum'
      useCount:
        example: 1
        type: integer
    type: object
  connection.ConnectionPoolResponse:
    properties:
      connections:
        items:
          $ref: '#/definitions/connection.ConnectionInfoResponse'
        type: array
      poolSize:
        type: integer
    type: object
  connection.ConnectionRequest:
    properties:
      certificate:
        description: для certificate
        example: cert-abc-123
        type: string
      connectionType:
        allOf:
        - $ref: '#/definitions/models.ConnectionTypeEnum'
        example: password
      endpointURL:
        example: opc.tcp://KHRLLW_-340595:4840/HEIDENHAIN/NC
        type: string
      key:
        example: secret
        type: string
      manufacturer:
        example: Heidenhain
        type: string
      mode:
        allOf:
        - $ref: '#/definitions/connection.MessageSecurityModeEnum'
        description: OPC UA MessageSecurityMode
        example: SignAndEncrypt
      model:
        example: TNC640
        type: string
      password:
        description: для password
        example: secret
        type: string
      policy:
        allOf:
        - $ref: '#/definitions/connection.SecurityPolicyEnum'
        description: OPC UA SecurityPolicy
        example: Basic256Sha256
      username:
        description: для password
        example: client1
        type: string
    required:
    - connectionType
    - manufacturer
    - model
    type: object
  connection.ConnectionStatusEnum:
    enum:
    - OK
    - FAIL
    - NOT FOUND
    - HEALTHY
    - UNHEALTHY
    - CONNECTED
    - DISCONNECTED
    - ERROR
    - TIMEOUT
    - AUTH_FAILED
    type: string
    x-enum-comments:
      StatusAuthFailed: Ошибка аутентификации
      StatusConnected: Соединение установлено
      StatusDisconnected: Соединение разорвано
      StatusError: Ошибка соединения
      StatusHealthy: Соединение здорово и работает
      StatusTimeout: Таймаут соединения
      StatusUnhealthy: Соединение есть, но есть проблемы
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - Соединение здорово и работает
    - Соединение есть, но есть проблемы
    - Соединение установлено
    - Соединение разорвано
    - Ошибка соединения
    - Таймаут соединения
    - Ошибка аутентификации
    x-enum-varnames:
    - StatusOK
    - StatusFail
    - StatusNotFound
    - StatusHealthy
    - StatusUnhealthy
    - StatusConnected
    - StatusDisconnected
    - StatusError
    - StatusTimeout
    - StatusAuthFailed
  connection.DisconnectResponse:
    properties:
      disconnected:
        type: boolean
    required:
    - disconnected
    type: object
  connection.IDRequest:
    properties:
      ID:
        type: integer
    required:
    - ID
    type: object
  connection.IDResponse:
    properties:
      ID:
        type: integer
    required:
    - ID
    type: object
  connection.MessageSecurityModeEnum:
    enum:
    - None
    - Sign
    - SignAndEncrypt
    type: string
    x-enum-varnames:
    - ModeNone
    - ModeSign
    - ModeSignAndEncrypt
  connection.PollingResponse:
    properties:
      polled:
        type: boolean
    required:
    - polled
    type: object
  connection.SecurityPolicyEnum:
    enum:
    - None
    - Basic128Rsa15
    - Basic256
    - Basic256Sha256
    type: string
    x-enum-varnames:
    - PolicyNone
    - PolicyBasic128Rsa15
    - PolicyBasic256
    - PolicyBasic256Sha256
  connection.StartPollingRequest:
    properties:
      ID:
        type: integer
      timeout:
        type: integer
    required:
    - ID
    - timeout
    type: object
  models.ConnectionConfig:
    properties:
      config: {}
    type: object
  models.ConnectionTypeEnum:
    enum:
    - anonymous
    - password
    - certificate
    type: string
    x-enum-varnames:
    - ConnectionAnonymous
    - ConnectionPassword
    - ConnectionCertificate
  swagger.CheckConnectionResponse:
    properties:
      data:
        $ref: '#/definitions/connection.ConnectionInfoResponse'
      message:
        example: Successfully get connection info
        type: string
      status:
        example: ok
        type: string
      type:
        example: object
        type: string
    type: object
  swagger.DisconnectResponse:
    properties:
      data:
        $ref: '#/definitions/connection.DisconnectResponse'
      message:
        example: Successfully disconnected
        type: string
      status:
        example: ok
        type: string
      type:
        example: object
        type: string
    type: object
  swagger.GetConnectionPoolResponse:
    properties:
      data:
        $ref: '#/definitions/connection.ConnectionPoolResponse'
      message:
        example: Successfully get connection pool
        type: string
      status:
        example: ok
        type: string
      type:
        example: object
        type: string
    type: object
  swagger.IDResponse:
    properties:
      data:
        $ref: '#/definitions/connection.IDResponse'
      message:
        example: Successfully connected
        type: string
      status:
        example: ok
        type: string
      type:
        example: object
        type: string
    type: object
  swagger.IncorrectDataError:
    properties:
      response:
        properties:
          code:
            example: 400
            type: integer
          message:
            example: Некорректные данные
            type: string
        type: object
      status:
        example: IncorrectDataError
        type: string
    type: object
  swagger.IncorrectFormatError:
    properties:
      response:
        properties:
          code:
            example: 400
            type: integer
          message:
            example: Неверный формат запроса
            type: string
        type: object
      status:
        example: IncorrectFormatError
        type: string
    type: object
  swagger.InternalServerError:
    properties:
      response:
        properties:
          code:
            example: 500
            type: integer
          message:
            example: Внутренняя ошибка сервера
            type: string
        type: object
      status:
        example: InternalServerError
        type: string
    type: object
  swagger.NotFoundError:
    properties:
      response:
        properties:
          code:
            example: 404
            type: integer
          message:
            example: Данные не найдены
            type: string
        type: object
      status:
        example: NotFoundError
        type: string
    type: object
  swagger.PollingResponse:
    properties:
      data:
        $ref: '#/definitions/connection.PollingResponse'
      message:
        example: Polling started/stopped for machine
        type: string
      status:
        example: ok
        type: string
      type:
        example: object
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для работы с протоколом OPC UA
  title: OpcUaService API
  version: 1.0.0
paths:
  /api/v1/polling/start:
    get:
      description: Запускает опрос OPC UA для конкретного станка по UUID
      parameters:
      - description: ID станка и интервал опроса (мс)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connection.StartPollingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Мониторинг запущен
          schema:
            $ref: '#/definitions/swagger.PollingResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Запустить мониторинг станка
      tags:
      - Polling
  /api/v1/polling/stop:
    get:
      description: Останавливает опрос OPC UA для конкретного станка по ID
      parameters:
      - description: UUID станка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connection.IDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Мониторинг запущен
          schema:
            $ref: '#/definitions/swagger.PollingResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Остановить мониторинг станка
      tags:
      - Polling
  /connect:
    delete:
      consumes:
      - application/json
      description: Закрывает соединение OPC UA по ID
      parameters:
      - description: ID для отключения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connection.IDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное отключение
          schema:
            $ref: '#/definitions/swagger.DisconnectResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Отключение сессии
      tags:
      - Connection
    get:
      description: Возвращает список активных соединений в пуле OPC UA
      produces:
      - application/json
      responses:
        "200":
          description: Список активных соединений
          schema:
            $ref: '#/definitions/swagger.GetConnectionPoolResponse'
      summary: Получить пул соединений
      tags:
      - Connection
    post:
      consumes:
      - application/json
      description: 'Подключение клиента для мониторинга станка: анонимно, по паролю
        или по сертификату. Передаётся JSON с данными для входа.'
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connection.ConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное подключение
          schema:
            $ref: '#/definitions/swagger.IDResponse'
        "400":
          description: Неверный формат запроса или некорректные данные
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "401":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/swagger.IncorrectDataError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Подключение клиента
      tags:
      - Connection
  /connect/check:
    post:
      consumes:
      - application/json
      description: Проверяет состояние соединения OPC UA по ID
      parameters:
      - description: ID для проверки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/connection.CheckConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Информация о подключении
          schema:
            $ref: '#/definitions/swagger.CheckConnectionResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Проверка соединения
      tags:
      - Connection
swagger: "2.0"